For this commit:
We created the Script: Signal_to_noise_gold.R which, takes all the
regions and extracted reads of an experiment and intersects them to
create the gold set. Then given a gold set, and the sets of processed
reads it extract the reads (of each sample) that overlaps with the
gold set.

With this sets of reads, it generates the same summary_statistics
table and plots the Signal to Noise plot under two views:

1 - Using hexbin to attempt to understand the SNR of the ChIP-exo
samples faceted by the sample

2 - It colours them according to it forward strand ratio, in an
attempt to understand the interaction between both analysis.

Then it saves the filtered version of this plot (by number of unique
positions) into two pdf files